Preserve EU actor boundary narrowing

+++ src/hum_common.c
@@ -438,7 +438,7 @@ s32 func_0800E5F0(HumWork* work) {
 
             func_0801B008();
 #ifdef VERSION_EU
-            func_0801A8A4(&actor->unk_004, &actor->unk_008, work->unk_174 - 8, -16);
+            func_0801A8A4(&actor->unk_004, &actor->unk_008, (s16)(work->unk_174 - 8), -16);
 #endif
             func_0801B918(actor);
             arg2.unk_00 = actor->unk_004;
